system_design/python/src/auth/requirements.txt
astroid==2.15.6
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dill==0.3.7
Flask==2.3.3
Flask-MySQLdb==1.0.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
tomlkit==0.12.1
Werkzeug==2.3.7
wrapt==1.15.0